ditto of 1.1.7

+++ src/main/java/net/alexandra/atlas/atlas_combat/mixin/RamTargetMixin.java
@@ -11,6 +11,6 @@ import org.spongepowered.asm.mixin.injection.Redirect;
 public class RamTargetMixin {
 	@Redirect(method = "tick(Lnet/minecraft/server/level/ServerLevel;Lnet/minecraft/world/entity/animal/goat/Goat;J)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;knockback(DDD)V"))
 	public void redirectKnockback(LivingEntity instance, double strength, double x, double z) {
-		((LivingEntityExtensions)instance).newKnockback((float)strength, x, z);
+		((LivingEntityExtensions)instance).newKnockback(strength, x, z);
 	}
 }
